end/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
mysqlclient==2.2.6
django-cors-headers==4.6.0
python-dotenv==1.0.1
